Changed all float32 dtypes to float64.
git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pysynphot@331 90a0a646-be8a-0410-bb88-9290da87bc01

+++ lib/catalog.py
@@ -71,7 +71,7 @@ class Icat(spectrum.TabularSourceSpectrum):
     def _breakList(self, inList, index, parameter):
         par = float(parameter)
 
-        array = N.empty(shape=[len(inList),],dtype=N.float32)
+        array = N.empty(shape=[len(inList),],dtype=N.float64)
         i = 0
         for parameters in inList:
             array[i] = parameters[index]
